fix(auth): resolve Flask 3 OIDC callback route AssertionError
In Flask 3 (v0.5.1+), dynamically calling @current_app.route(...) after the application has started handling requests raises an AssertionError. Removes dynamic route registration from inside oidc_oauth() and statically registers @index_bp.route('/oidc/authorized') on index_bp in routes/index.py, updating url_for references to use the index.oidc_authorized blueprint endpoint.
